remove SourceUtil
